노트스피드 정수로 사용

+++ Term_Project/Term_Project/option_state.h
@@ -79,7 +79,7 @@ public:
 	virtual void InitBuffer();
 
 	// system
-	GLfloat Get_Note_Speed();
+	GLint Get_Note_Speed();
 	GLint Get_Volume();
 };
